wip(#1099): extract lower_call/buffer_intrinsic
